chore:  modify CCDS .gitignore to include data directory
The CCDS Opinions documentation advises excluding datasets from source
control: "Another consequence of treating data as immutable is that data
doesn't need source control in the same way that code does. Therefore,
by default, the data/ folder is included in the .gitignore file."
However, for this repository, the dataset is included in the data
directory to facilitate learning and ensure ease of access during the
project's development.